Fixes an issue with Jackson overriding camelCase to snake case even when using different configuration within the feign client, fixed by defining a decoder and encoder ObjectMapper and adding it as configuration to the feign client, this was causing issues when retrieving search results from camunda eventually returning everything as it was ignoring the other properties.
Added new structure to the search query and added an operator property which would be used in the future to define the explicitly operations and allows for better future improvements.

Removing ccdId eventId pre event state and post event state from the queries as these will be covered by another ticket.

Minor fixes to default constructors.

Added UnsupportedOperationException in callback controller to map to a 400, this is useful when using operators that we do not  yet support