Add uniform role assignment signatures.
Using different schemes for organisational and case role assignments
may lead to problems in the future if patterns of role assignment usage
change, even though it reduces the size of the GIN index.

This changes reverts to the original signature generation mechanism,
which is the same for all role assignments.

The change is enabled / disabled in TaskPerfConfig and using database
configuration parameters.