Stop expanding classifications in role assignment signatures.
It is necessary to match all tasks with a classification <= the role assignment.
Currently, the GIN index expands the signatures for all classifications >= the task
classification, AND the signatures generated for role assignments include all
classifications >= the role assignment classification.
We don't need to do both, so the latter has been removed.