Rwa-1437-3 (#789)
* Add uniform role assignment signatures.

Using different schemes for organisational and case role assignments
may lead to problems in the future if patterns of role assignment usage
change, even though it reduces the size of the GIN index.

This changes reverts to the original signature generation mechanism,
which is the same for all role assignments.

The change is enabled / disabled in TaskPerfConfig and using database
configuration parameters.

* Do not filter out jurisdiction-less role assignments.

The implementation filters out role assignments which cannot be used for
tasks.  This has included any role assignment without a jurisdiction.
However, that may not be correct long-term, especially if wide-ranging
cross-jurisdiction roles are added (e.g. GDPR staff).

* Stop expanding classifications in role assignment signatures.

It is necessary to match all tasks with a classification <= the role assignment.
Currently, the GIN index expands the signatures for all classifications >= the task
classification, AND the signatures generated for role assignments include all
classifications >= the role assignment classification.
We don't need to do both, so the latter has been removed.

* Fix potential security hole with conflicting constraints.

See the comment in ClientFilter.

* Fix bug in filter signature generation with null task attributes.

For task attributes which are null, signatures should be generated
using only a wildcard.  The implementation was including the null value,
leading to lots of null signatures being generated.

* Turn on uniform role assignment signatures by default.

* Change SQL for filter signatures to the same style as role signatures.

The role signature SQL was easier to read and understand, so the filter
signature SQL is updated to follow the same pattern.

* Bumping chart version/ fixing aliases

Co-authored-by: Jon Moore <jonathan.moore@hmcts.net>
Co-authored-by: hmcts-jenkins-j-to-z <61242337+hmcts-jenkins-j-to-z[bot]@users.noreply.github.com>